requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
blinker==1.4
certifi==2017.7.27.1      # via requests
chardet==3.0.4            # via requests
click==6.7
idna==2.6                 # via requests
msgpack-python==0.4.8
nose==1.3.7               # via screepsapi
pluginbase==0.5
pydevd==1.1.1
requests==2.18.4
requires==0.0.3           # via screepsapi
screepsapi==0.4.5
six==1.11.0               # via websocket-client
urllib3==1.22             # via requests
urwid==1.3.1
websocket-client==0.44.0  # via screepsapi
